% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{GetClientBalance}
\alias{GetClientBalance}
\title{Get Client Balance Response}
\usage{
GetClientBalance(force = FALSE)
}
\arguments{
\item{force}{boolean if TRUE relaod the data from the site otherwise use the cache}
}
\value{
anmed vector client balance parameter
}
\description{
Get Client Balance Response
}
\examples{
GetClientBalance()
}

