% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{GetCurrencies}
\alias{GetCurrencies}
\title{Get the list of supported currencies}
\usage{
GetCurrencies(force = FALSE)
}
\arguments{
\item{force}{boolean if false use cached data.}
}
\value{
a data frame having columns:
\itemize{
\item Code
\item Rate
\item Name
}
}
\description{
Get the list of supported currencies
}
\examples{
GetCurrencies()
}

