% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{GetLeaguesByID}
\alias{GetLeaguesByID}
\title{Get leagues for a sportid}
\usage{
GetLeaguesByID(sportid)
}
\arguments{
\item{sportid}{integer the id of the sport}
}
\value{
a data frame having columns:
\itemize{
\item League ID
\item  Feed Contents
\item HomeTeamType
\item AllowRoundRobin
\item League Name
}
}
\description{
retrieves all sports leagues with the status whether they currently have lines or not.
}

