% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{GetOdds}
\alias{GetOdds}
\title{Get Odds}
\usage{
GetOdds(sportname, leagueids, since, isLive)
}
\arguments{
\item{sportname}{The sport name for which to retrieve the fixutres}

\item{leagueids}{vector of characters.}

\item{since}{numeric This is used to receive incremental updates.
Use the value of last from previous fixtures response.}

\item{isLive}{boolean if TRUE retrieves ONLY live events}
}
\value{
list of lists
}
\description{
Get Odds
}
\examples{
GetOdds (sportname="Soccer", leagueid=c(11,45),since=26142345,isLive=0)
}

