% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{GetSports}
\alias{GetSports}
\title{Get Sports}
\usage{
GetSports(force)
}
\arguments{
\item{force}{boolean if false use cached data.}
}
\value{
a data frame having columns:
\itemize{
\item Sport ID
\item  Feed Contents
\item Sport Name
}
}
\description{
Returns all sports with the status whether they currently have lines or not
}
\examples{
GetSports()
}

