% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{PlaceBet}
\alias{PlaceBet}
\title{Place Bet}
\usage{
PlaceBet(acceptBetterLine, oddsFormat, stake, winRiskStake, sportId, eventId,
  periodNumber, betType, lineId)
}
\arguments{
\item{acceptBetterLine}{boolean Whether or not to accept a bet
when there is a line change in favor of the client}

\item{oddsFormat}{character Bet is processed with this odds format
it takes values in :
\itemize{
\item{AMERICAN}
\item{DECIMAL}
\item{HONGKONG}
\item{INDONESIAN}
\item{MALAY}
}}

\item{stake}{numeric Wagered amount in Client’s currency}

\item{winRiskStake}{WIN_RISK_TYPE Whether the stake amount is risk or win amount
\itemize{
\item{WIN}
\item{RISK}
}}

\item{sportId}{numeric the sport id}

\item{eventId}{numeric the vent id}

\item{periodNumber}{This represents the period of the match. For example
For soccer : #' 0-->Game 1-->1st Half 2-->2nd Half}

\item{betType}{BET_TYPE
\itemize{
\item{SPREAD    }
\item{MONEYLINE}
\item{TOTAL_POINTS}
\item{TEAM_TOTAL_POINTS}
}}

\item{lineId}{numeric Line identification}
}
\value{
list containing :
\itemize{
  \item{status}  If Status is PROCESSED_WITH_ERROR errorCode will be in the response
  \item{errorCode}
}
}
\description{
Place bet in the system
}
\examples{
PlaceBet (acceptBetterLine=TRUE,oddsFormat="AMERICAN", stake=10,
          winRiskStake="WIN",sportId=29,eventId=307962592,
           periodNumber=0,betType="MONEYLINE",lineId=103648474)
}

